+++ src/lib/pki/organisation.spec.ts
@@ -1,7 +1,7 @@
 import { jest } from '@jest/globals';
 import { addMinutes, setMilliseconds, subMinutes } from 'date-fns';
 
-import { derSerializePublicKey } from '../utils/keys.js';
+import { derSerializePublicKey } from '../utils/keys/serialisation.js';
 import type FullIssuanceOptions from '../utils/x509/FullIssuanceOptions.js';
 import Certificate from '../utils/x509/Certificate.js';
 import { getBasicConstraintsExtension } from '../../testUtils/pkijs.js';
