+++ src/testUtils/pki.ts
@@ -1,8 +1,9 @@
 import { addDays, setMilliseconds } from 'date-fns';
 
 import Certificate from '../lib/utils/x509/Certificate.js';
-import { derSerializePublicKey, generateRsaKeyPair } from '../lib/utils/keys.js';
+import { derSerializePublicKey } from '../lib/utils/keys/serialisation.js';
 import type FullIssuanceOptions from '../lib/utils/x509/FullIssuanceOptions.js';
+import { generateRsaKeyPair } from '../lib/utils/keys/generation.js';
 
 import { reSerializeCertificate } from './pkijs.js';
 import { calculateDigest } from './crypto.js';
