+++ src/testUtils/veraStubs/member.ts
@@ -1,4 +1,4 @@
-import { generateRsaKeyPair } from '../../lib/utils/keys.js';
+import { generateRsaKeyPair } from '../../lib/utils/keys/generation.js';
 
 export const MEMBER_NAME = 'alice';
 export const MEMBER_KEY_PAIR = await generateRsaKeyPair();
